requirements.txt
streamlit==1.35.0
openai==1.14.2
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
